+++ CHANGELOG.md
@@ -6,8 +6,9 @@ PushmiPullyu is a Ruby application, whose primary job is to manage the flow of c
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and releases in PushmiPullyu adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.1.0]
  - Add more logging information [#433](https://github.com/ualbertalib/pushmi_pullyu/issues/433)
+ - Add V3 authentication [#349](https://github.com/ualbertalib/pushmi_pullyu/issues/349)
 
 ## [2.0.7] - 2023-09-13